pyproject.toml
[project]
name = "ai-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.12",
    "google-generativeai==0.8.5",
    "openai>=1.87.0",
    "pydantic==2.11.5",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]==3.5.0",
    "uvicorn==0.34.3",
]
